Refactor MetricsAndLoggingUtils to properly separate Model and Operator classes
