+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/MetricsAndLoggingUtils.java
@@ -2,23 +2,18 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.cluster.model;
+package io.strimzi.operator.cluster.operator.assembly;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.strimzi.api.kafka.model.ExternalLogging;
 import io.strimzi.operator.cluster.model.logging.LoggingModel;
-import io.strimzi.operator.cluster.model.logging.SupportsLogging;
 import io.strimzi.operator.cluster.model.metrics.MetricsModel;
-import io.strimzi.operator.cluster.model.metrics.SupportsMetrics;
-import io.strimzi.operator.common.MetricsAndLogging;
+import io.strimzi.operator.cluster.model.MetricsAndLogging;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.model.InvalidResourceException;
 import io.strimzi.operator.common.operator.resource.ConfigMapOperator;
 import io.vertx.core.Future;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * Shared methods for working with Metrics and Logging configurations. These methods are bundled because we store both
@@ -69,31 +64,4 @@ public class MetricsAndLoggingUtils {
         }
     }
 
-    /**
-     * Generates a metrics and logging ConfigMap according to configured defaults. This is used with most operands, but
-     * not all of them. Kafka brokers have own methods in the KafkaCluster class. So does the Bridge. And Kafka Exporter
-     * has no metrics or logging ConfigMap at all.
-     *
-     * @param reconciliation        Reconciliation marker
-     * @param model                 AbstractModel instance possibly implementing SupportsLogging and SupportsMetrics interfaces
-     * @param metricsAndLogging     ConfigMaps with external logging and metrics configuration
-     *
-     * @return The generated ConfigMap.
-     */
-    public static Map<String, String> generateMetricsAndLogConfigMapData(Reconciliation reconciliation, AbstractModel model, MetricsAndLogging metricsAndLogging) {
-        Map<String, String> data = new HashMap<>(2);
-
-        if (model instanceof SupportsLogging supportsLogging) {
-            data.put(supportsLogging.logging().configMapKey(), supportsLogging.logging().loggingConfiguration(reconciliation, metricsAndLogging.loggingCm()));
-        }
-
-        if (model instanceof SupportsMetrics supportMetrics) {
-            String parseResult = supportMetrics.metrics().metricsJson(reconciliation, metricsAndLogging.metricsCm());
-            if (parseResult != null) {
-                data.put(MetricsModel.CONFIG_MAP_KEY, parseResult);
-            }
-        }
-
-        return data;
-    }
 }
